+++ legacy.js
@@ -8,14 +8,16 @@ module.exports = {
     './rules/style',
     './rules/variables'
   ].map(require.resolve),
-  'env': {
-    'browser': true,
-    'node': true,
-    'amd': false,
-    'mocha': false,
-    'jasmine': false
+  env: {
+    browser: true,
+    node: true,
+    amd: false,
+    mocha: false,
+    jasmine: false
   },
-  'ecmaFeatures': {},
-  'globals': {},
-  'rules': {}
+  parserOptions: {
+    ecmaFeatures: {}
+  },
+  globals: {},
+  rules: {}
 };